+++ e2e/wdio/wdio.local.conf.ts
@@ -28,6 +28,11 @@ export const config: Options.Testrunner = {
         'moz:firefoxOptions': {
             args: ['-headless']
         }
+    }, {
+        browserName: 'edge',
+        'ms:edgeOptions': {
+            args: ['headless', 'disable-gpu']
+        }
     }],
 
     /**
@@ -47,14 +52,5 @@ export const config: Options.Testrunner = {
 if (os.platform() === 'darwin') {
     (config.capabilities as Capabilities.Capabilities[]).push({
         browserName: 'safari'
-    })
-}
-
-if (['darwin', 'win32'].includes(os.platform())) {
-    (config.capabilities as Capabilities.Capabilities[]).push({
-        browserName: 'edge',
-        'ms:edgeOptions': {
-            args: ['--headless', '--disable-gpu']
-        }
     })
 }