+++ packages/wdio-utils/tests/driver/index.test.ts
@@ -24,6 +24,13 @@ vi.mock('node:fs', () => ({
     }
 }))
 
+vi.mock('node:os', async (origMod) => ({
+    default: {
+        ...(await origMod<any>()),
+        platform: vi.fn().mockReturnValue('linux')
+    }
+}))
+
 vi.mock('node:child_process', () => ({
     default: {
         spawn: vi.fn().mockReturnValue({
@@ -38,7 +45,10 @@ vi.mock('@wdio/logger', () => import(path.join(process.cwd(), '__mocks__', '@wdi
 vi.mock('devtools', () => ({ default: 'devtools package' }))
 vi.mock('webdriver', () => ({ default: 'webdriver package' }))
 vi.mock('safaridriver', () => ({ start: vi.fn().mockReturnValue('safaridriver') }))
-vi.mock('edgedriver', () => ({ start: vi.fn().mockResolvedValue('edgedriver') }))
+vi.mock('edgedriver', () => ({
+    start: vi.fn().mockResolvedValue('edgedriver'),
+    findEdgePath: vi.fn().mockReturnValue('/foo/bar/executable')
+}))
 vi.mock('geckodriver', () => ({ start: vi.fn().mockResolvedValue('geckodriver') }))
 vi.mock('wait-port', () => ({ default: vi.fn() }))
 vi.mock('get-port', () => ({ default: vi.fn().mockResolvedValue(1234) }))
@@ -142,6 +152,9 @@ describe('startWebDriver', () => {
                 'wdio:edgedriverOptions': {
                     foo: 'bar'
                 },
+                'ms:edgeOptions': {
+                    binary: '/foo/bar/executable'
+                }
             }
         })
         expect(startEdgedriver).toBeCalledTimes(1)