+++ packages/webdriver/src/index.ts
@@ -1,12 +1,11 @@
 import path from 'node:path'
 import logger from '@wdio/logger'
 
-import { webdriverMonad, sessionEnvironmentDetector } from '@wdio/utils'
+import { webdriverMonad, sessionEnvironmentDetector, startWebDriver } from '@wdio/utils'
 import { validateConfig } from '@wdio/config'
 import type { Options, Capabilities } from '@wdio/types'
 
 import command from './command.js'
-import { startWebDriver } from './driver/index.js'
 import { BidiHandler } from './bidi/handler.js'
 import { DEFAULTS } from './constants.js'
 import { startWebDriverSession, getPrototype, getEnvironmentVars, setupDirectConnect } from './utils.js'
@@ -37,8 +36,8 @@ export default class WebDriver {
         }
 
         log.info('Initiate new session using the WebDriver protocol')
-        const requestedCapabilities = { ...params.capabilities }
         const driverProcess = await startWebDriver(params)
+        const requestedCapabilities = { ...params.capabilities }
         const { sessionId, capabilities } = await startWebDriverSession(params)
         const environment = sessionEnvironmentDetector({ capabilities, requestedCapabilities })
         const environmentPrototype = getEnvironmentVars(environment)